crm114-20080502-BlameJason-Ger-2257
version 20080502-BlameJason

 * multiple out-of-bounds security issues fixed

 * extended CRM114 debugger funtionality - see QUICKREF.txt

 * 'make check' now performs all megatest tests; individual tests can be
   run separately - see tests/Makefile.am for the make target names

 * more error condition checking and reporting than ever before, also
   the compiler now perform several table-driven checks (TODO: argument
   count verification).

   So invalid scripts may no longer be silently accepted by CRM114.

 * megatest_knowngood.log split up across 'master reference files': one
   for each of the tests in tests/

 * when the script may/can/will be debugged, an extra script variable
   will exist:

     :_?:

   which can be watched in the debugger, e.g. by specifying the watch
   var select regex

     + _\?

   This variable will store the last error message that was generated by
   CRM114 while running the script.

 * geekiness: upgraded the build environment for maintainers to the
   latest autoconf/make/etc. tool set

 KNOWN BUGS:

 - experimental classifiers are generally a mess.

 - some tests do not pass (see tests/Makefile.am: search for OVERRIDE

 - syscall on Win32 is still ever so slightly b0rked - next release
   should provide a completely rewritten code section for that.

+++ src/crm_main.c
@@ -191,6 +191,7 @@ void free_stack(CSL_CELL *csl)
  *    so tread carefully here: do not assume all these pointers are filled.
  */
 static void crm_final_cleanup(void)
+
 {
     // GROT GROT GROT
     //
@@ -1513,7 +1514,7 @@ int main(int argc, char **argv)
         int dwlen;
         tdw->filetext[tdw->nchars] = '\n';
         tdw->nchars++;
-        dwlen = strlen(":_dw:");
+        dwlen = (int)strlen(":_dw:");
         dwname = tdw->nchars;
         //strcat (tdw->filetext, ":_dw:");
         memmove(&tdw->filetext[dwname], ":_dw:", dwlen);
